+++ src/snovault/tests/test_snowflake_hash.py
@@ -9,5 +9,5 @@ TEST_HASHES = {
 
 @pytest.mark.parametrize(('password', 'pwhash'), TEST_HASHES.items())
 def test_snowflake_hash(password, pwhash):
-    from snovault.snowflake_hash import SNOWHash
+    from snovault.tests.snowflake_hash import SNOWHash
     assert SNOWHash.encrypt(password) == pwhash
